+++ demisto_sdk/common/hook_validations/image.py
@@ -1,8 +1,9 @@
 import glob
 import base64
 
-from demisto_sdk.common.tools import re, print_error, os, get_yaml
-from demisto_sdk.common.constants import IMAGE_REGEX, INTEGRATION_REGEX, INTEGRATION_YML_REGEX, DEFAULT_IMAGE_BASE64
+from demisto_sdk.common.tools import re, print_error, os, get_yaml, checked_type
+from demisto_sdk.common.constants import IMAGE_REGEX, INTEGRATION_REGEX, DEFAULT_IMAGE_BASE64, \
+    INTEGRATION_REGXES, YML_INTEGRATION_REGEXES
 
 
 class ImageValidator:
@@ -17,10 +18,10 @@ class ImageValidator:
     def __init__(self, file_path):
         self._is_valid = True
 
-        if re.match(INTEGRATION_REGEX, file_path, re.IGNORECASE):
+        if checked_type(file_path, INTEGRATION_REGXES):
             self.file_path = file_path
         else:
-            if re.match(INTEGRATION_YML_REGEX, file_path, re.IGNORECASE):
+            if checked_type(file_path, YML_INTEGRATION_REGEXES):
                 try:
                     self.file_path = glob.glob(os.path.join(os.path.dirname(file_path), '*.png'))[0]
                 except IndexError: