+++ CHANGELOG.md
@@ -7,7 +7,44 @@
 * Added spell check command.
 
 
+### 0.2.6
+
+* Fixed an issue with locating release notes for beta integrations in **validate**.
+
+### 0.2.5
+
+* Fixed an issue with locating release notes for beta integrations in **validate**.
+
+### 0.2.4
+
+* Adding image validation to Beta_Integration and Packs in **validate**.
+
+### 0.2.3
+
+* Adding Beta_Integration to the structure validation process.
+* Fixing bug where **validate** did checks on TestPlaybooks.
+* Added requirements parameter to **lint**.
+
+### 0.2.2
+
+* Fixing bug where **lint** did not return exit code 1 on failure.
+* Fixing bug where **validate** did not print error message in case no release notes were give.
+
+### 0.2.1
+
+* **Validate** now checks that the id and name fields are identical in yml files.
+* Fixed a bug where sdk did not return any exit code.
+
+### 0.2.0
+
+* Added Release Notes Validator.
+* Fixed the Unifier selection of your python file to use as the code.
+* **Validate** now supports Indicator fields.
+* Fixed a bug where **validate** and **secrets** did not return exit code 1 on failure.
+* **Validate** now runs on newly added scripts.
+
 ### 0.1.8
+
 * Added support for `--version`.
 * Fixed an issue in file_validator when calling `checked_type` method with script regex.